Create full-fledged viewer prototype
Creating a prototype of the viewer with the following functionality:
- Browse pages in coverflow, load and highlight clicked page
- Only load coverflow thumbs when they become visible, enable scrolling with mousewheel
- Toggle between single and double page display
- Previous and next buttons
- Zoom and pan image
- Toggle transcription, TOC and metadata panes
- Toggle fullscreen

A new fixed-layout template was added along with multiple partials and styles. Translations were updated. Fotorama was removed.

The `fonts` build task was replaced by an `assets` tasks, handling all files that are copied over to `/web` unchanged.

Known issues:
- GUI elements not mentioned above are inoperable yet
- URLs ending in "|LOG_nnnn" do not lead to the correct page
- Zooming double pages has transform origin issues
- There may be missing translations

Install new dependencies with `npm install`.