Make image cache a little more persistent
The image cache is emptied each time the application cache is cleared.
When heaving continuous deploys, this may slow down the app, as the
images always need to be recalculated. This adds a new image cache in
var/images/ which is not cleared automatically.
A new command needs to be implemented to provide an image-cache-clearing
feature.