Add basic handling for anchor documents
This adds a condition if a document is an anchor document and does not
contain contents itself.
A call is done to get all documents with a parent id of that anchor
document, reduced to the documents where only one parent is existing
(for now) and displayed and linked in the frontend.